Create Orange Squadron Communication Code
This code takes place within our condition that our robot has scanned a black line.  We have a counter to detect the hash marks that we need to scan colors, ie the counter variable.  Once that is done, ie line 125, it has the conditions on that line.  It stops and continuously scans for the incoming number.  Once that condition is met, ie line 134, our bot gets to the desired line. When it gets there, ie line 139, it stops, outputs its number once, and keeps hitting that loop in the void loop forever.  Boom.